refactor: flatten src layout
