+++ battle_agent_random/tests/test_randomplayer.py
@@ -1,11 +1,11 @@
 import unittest
 import uuid
-from battle_agent_random.randomplayer import RandomPlayer
-from battle_hexes_core.game.board import Board
-from battle_hexes_core.game.game import Game
-from battle_hexes_core.game.player import Player, PlayerType
-from battle_hexes_core.unit.faction import Faction
-from battle_hexes_core.unit.unit import Unit
+from randomplayer import RandomPlayer
+from game.board import Board
+from game.game import Game
+from game.player import Player, PlayerType
+from unit.faction import Faction
+from unit.unit import Unit
 import random
