+++ battle_hexes_api/src/game_factory.py
@@ -1,11 +1,11 @@
 from uuid import UUID
 
-from battle_agent_random.randomplayer import RandomPlayer
-from battle_hexes_core.game.board import Board
-from battle_hexes_core.game.game import Game
-from battle_hexes_core.game.player import PlayerType
-from battle_hexes_core.unit.faction import Faction
-from battle_hexes_core.unit.unit import Unit
+from randomplayer import RandomPlayer
+from game.board import Board
+from game.game import Game
+from game.player import PlayerType
+from unit.faction import Faction
+from unit.unit import Unit
 
 
 class GameFactory:
