+++ battle_hexes_api/src/main.py
@@ -15,9 +15,9 @@ sys.path.insert(0, str(REPO_ROOT / "battle_agent_random" / "src"))
 sys.path.insert(0, str(REPO_ROOT / "battle_agent_rl" / "src"))
 print("sys.path:", sys.path)
 
-from battle_hexes_core.combat.combat import Combat  # noqa: E402
-from battle_hexes_core.game.gamerepo import GameRepository  # noqa: E402
-from battle_hexes_core.game.sparseboard import SparseBoard  # noqa: E402
+from combat.combat import Combat  # noqa: E402
+from game.gamerepo import GameRepository  # noqa: E402
+from game.sparseboard import SparseBoard  # noqa: E402
 from src.game_factory import GameFactory  # noqa: E402
 
 app = FastAPI()
