+++ battle_hexes_core/src/combat/combat.py
@@ -1,10 +1,10 @@
-from battle_hexes_core.game.game import Game
-from battle_hexes_core.combat.combatresult import (
+from game.game import Game
+from combat.combatresult import (
     CombatResult,
     CombatResultData,
 )
-from battle_hexes_core.combat.combatresults import CombatResults
-from battle_hexes_core.combat.combatsolver import CombatSolver
+from combat.combatresults import CombatResults
+from combat.combatsolver import CombatSolver
 
 
 class Combat:
