+++ battle_hexes_core/src/combat/combatresult.py
