+++ battle_hexes_core/src/game/hex.py
