+++ battle_hexes_core/src/unit/faction.py
