+++ battle_hexes_core/src/unit/sparseunit.py
