+++ battle_hexes_core/tests/game/test_board.py
@@ -1,10 +1,10 @@
 import unittest
 import uuid
-from battle_hexes_core.game.board import Board
-from battle_hexes_core.game.player import Player, PlayerType
-from battle_hexes_core.game.sparseboard import SparseBoard
-from battle_hexes_core.unit.faction import Faction
-from battle_hexes_core.unit.unit import Unit
+from game.board import Board
+from game.player import Player, PlayerType
+from game.sparseboard import SparseBoard
+from unit.faction import Faction
+from unit.unit import Unit
 
 
 class TestBoard(unittest.TestCase):
