+++ battle_hexes_core/tests/game/test_gamerepo.py
@@ -1,7 +1,7 @@
 import unittest
-from battle_hexes_core.game.board import Board
-from battle_hexes_core.game.game import Game
-from battle_hexes_core.game.gamerepo import GameRepository
+from game.board import Board
+from game.game import Game
+from game.gamerepo import GameRepository
 
 
 class TestGameRepos(unittest.TestCase):
